pyproject.toml
[tool.poetry]
name = "telegram-auth"
version = "0.1.0"
description = ""
authors = ["EgorLozh <lozhkinyegor@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.4"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
pydantic-settings = "^2.7.0"
punq = "^0.7.0"
aiogram = "^3.15.0"
django-rest-framework = "^0.1.0"
requests = "^2.32.3"
celery = {extras = ["redis"], version = "^5.4.0"}
django-celery-results = "^2.5.1"
pika = "^1.3.2"
pytz = "^2024.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
